remove httpUtils to common
